requirements.txt
langchain==0.3.25
langgraph==0.4.3
apscheduler==3.10.4
python-dotenv==1.0.0
openai==1.6.1
langchain_openai==0.0.8
requests==2.31.0
fastapi==0.103.2
uvicorn==0.23.2
